+++ tests/unit/test_core/conftest.py
@@ -2,6 +2,11 @@ import pytest
 
 from wakepy.core.method import Method, PlatformName
 
+# B, D, E
+FIRST_MODE = "first_mode"
+# A, F
+SECOND_MODE = "second_mode"
+
 
 @pytest.fixture(scope="function")
 def provide_methods_different_platforms(monkeypatch):
@@ -39,3 +44,32 @@ def provide_methods_different_platforms(monkeypatch):
             PlatformName.WINDOWS,
             PlatformName.MACOS,
         )
+
+
+@pytest.fixture(scope="function")
+def provide_methods_a_f(monkeypatch):
+    # empty method registry
+    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
+
+    class MethodA(Method):
+        name = "A"
+        mode = SECOND_MODE
+
+    class MethodB(Method):
+        name = "B"
+        mode = FIRST_MODE
+
+    class MethodC(Method):
+        name = "C"
+
+    class MethodD(Method):
+        name = "D"
+        mode = FIRST_MODE
+
+    class MethodE(Method):
+        name = "E"
+        mode = FIRST_MODE
+
+    class MethodF(Method):
+        name = "F"
+        mode = SECOND_MODE
