+++ tests/unit/test_core/test_method.py
@@ -5,51 +5,12 @@ import pytest
 from wakepy.core.method import (
     Method,
     MethodDefinitionError,
-    PlatformName,
-    check_methods_priority,
     get_method,
     get_methods,
     get_methods_for_mode,
-    get_prioritized_methods,
-    get_prioritized_methods_groups,
     method_names_to_classes,
     select_methods,
-    sort_methods_by_priority,
 )
-
-# B, D, E
-FIRST_MODE = "first_mode"
-# A, F
-SECOND_MODE = "second_mode"
-
-
-@pytest.fixture(scope="function")
-def provide_methods_a_f(monkeypatch):
-    # empty method registry
-    monkeypatch.setattr("wakepy.core.method._method_registry", dict())
-
-    class MethodA(Method):
-        name = "A"
-        mode = SECOND_MODE
-
-    class MethodB(Method):
-        name = "B"
-        mode = FIRST_MODE
-
-    class MethodC(Method):
-        name = "C"
-
-    class MethodD(Method):
-        name = "D"
-        mode = FIRST_MODE
-
-    class MethodE(Method):
-        name = "E"
-        mode = FIRST_MODE
-
-    class MethodF(Method):
-        name = "F"
-        mode = SECOND_MODE
 
 
 def test_overridden_methods_autodiscovery():
@@ -221,12 +182,12 @@ def test_get_methods_for_mode():
     methods = get_methods(["A", "B", "C", "D", "E", "F"])
     (MethodA, MethodB, _, MethodD, MethodE, MethodF) = methods
 
-    assert get_methods_for_mode(FIRST_MODE) == [
+    assert get_methods_for_mode("first_mode") == [
         MethodB,
         MethodD,
         MethodE,
     ]
-    assert get_methods_for_mode(SECOND_MODE) == [
+    assert get_methods_for_mode("second_mode") == [
         MethodA,
         MethodF,
     ]
@@ -256,221 +217,3 @@ def test_select_methods():
         ),
     ):
         select_methods(methods, use_only=["foo", "bar"])
-
-
-@pytest.mark.usefixtures("provide_methods_a_f")
-def test_check_methods_priority():
-    methods = get_methods(["A", "B", "C", "D", "E", "F"])
-    (MethodA, *_) = methods
-
-    # These should be fine
-    check_methods_priority(methods_priority=None, methods=methods)
-    check_methods_priority(methods_priority=["*"], methods=methods)
-    # Simple list of methods
-    check_methods_priority(methods_priority=["A", "B", "F"], methods=methods)
-    # Simple list of methods with asterisk
-    check_methods_priority(methods_priority=["A", "B", "*", "F"], methods=methods)
-    # Simple set + strings
-    check_methods_priority(methods_priority=[{"A", "B"}, "*", "F"], methods=methods)
-    # Simple set + strings
-    check_methods_priority(
-        methods_priority=[{"A", "B"}, "*", "E", {"F"}], methods=methods
-    )
-
-    # These are not fine
-    # There is no Method with name "X" in methods
-    with pytest.raises(
-        ValueError,
-        match=re.escape('Method "X" in methods_priority not in selected methods!'),
-    ):
-        check_methods_priority(methods_priority=["X"], methods=methods)
-
-    # two asterisks
-    with pytest.raises(
-        ValueError,
-        match=re.escape("The asterisk (*) can only occur once in methods_priority!"),
-    ):
-        check_methods_priority(methods_priority=["A", "*", "B", "*"], methods=methods)
-
-    # Asterisk inside a set
-    with pytest.raises(
-        ValueError,
-        match=re.escape("Asterisk (*) may not be a part of a set in methods_priority!"),
-    ):
-        check_methods_priority(methods_priority=[{"*"}], methods=methods)
-
-    # Unsupported type
-    with pytest.raises(
-        TypeError,
-        match=re.escape("methods_priority must be a list[str | set[str]]!"),
-    ):
-        check_methods_priority(methods_priority=[MethodA], methods=methods)
-
-
-@pytest.mark.usefixtures("provide_methods_a_f")
-def test_get_prioritized_methods_groups_does_not_edit_args():
-    """Test that the prioriry_order argument is not modified by the function"""
-    methods = get_methods(["A", "B", "C", "D", "E", "F"])
-
-    methods_priority = ["A", "F"]
-
-    _ = get_prioritized_methods_groups(
-        methods,
-        methods_priority=methods_priority,
-    )
-
-    assert methods_priority == [
-        "A",
-        "F",
-    ], "The methods_priority argument should not be modified by the function"
-
-
-@pytest.mark.usefixtures("provide_methods_a_f")
-def test_get_prioritized_methods_groups():
-    methods = get_methods(["A", "B", "C", "D", "E", "F"])
-    (MethodA, MethodB, MethodC, MethodD, MethodE, MethodF) = methods
-
-    # Case: Select some methods as more important, with '*'
-    assert get_prioritized_methods_groups(
-        methods, methods_priority=["A", "F", "*"]
-    ) == [
-        {MethodA},
-        {MethodF},
-        {MethodB, MethodC, MethodD, MethodE},
-    ]
-
-    # Case: Select some methods as more important, without '*'
-    # The results should be exactly the same as with asterisk in the end
-    assert get_prioritized_methods_groups(
-        methods,
-        methods_priority=["A", "F"],
-    ) == [
-        {MethodA},
-        {MethodF},
-        {MethodB, MethodC, MethodD, MethodE},
-    ]
-
-    # Case: asterisk in the middle
-    assert get_prioritized_methods_groups(
-        methods, methods_priority=["A", "*", "B"]
-    ) == [
-        {MethodA},
-        {MethodC, MethodD, MethodE, MethodF},
-        {MethodB},
-    ]
-
-    # Case: asterisk at the start
-    assert get_prioritized_methods_groups(
-        methods, methods_priority=["*", "A", "B"]
-    ) == [
-        {MethodC, MethodD, MethodE, MethodF},
-        {MethodA},
-        {MethodB},
-    ]
-
-    # Case: Using sets
-    assert get_prioritized_methods_groups(
-        methods, methods_priority=[{"A", "B"}, "*", {"E", "F"}]
-    ) == [
-        {MethodA, MethodB},
-        {MethodC, MethodD},
-        {MethodE, MethodF},
-    ]
-
-    # Case: Using sets, no asterisk -> implicit asterisk at the end
-    assert get_prioritized_methods_groups(methods, methods_priority=[{"A", "B"}]) == [
-        {MethodA, MethodB},
-        {MethodC, MethodD, MethodE, MethodF},
-    ]
-
-    # Case: methods_priority is None -> Should return all methods as one set
-    assert get_prioritized_methods_groups(methods, methods_priority=None) == [
-        {MethodA, MethodB, MethodC, MethodD, MethodE, MethodF},
-    ]
-
-
-@pytest.mark.usefixtures("provide_methods_different_platforms")
-def test_sort_methods_by_priority(monkeypatch):
-    WindowsA, WindowsB, WindowsC, LinuxA, LinuxB, LinuxC, MultiPlatformA = get_methods(
-        ["WinA", "WinB", "WinC", "LinuxA", "LinuxB", "LinuxC", "multiA"]
-    )
-
-    monkeypatch.setattr("wakepy.core.method.CURRENT_PLATFORM", PlatformName.LINUX)
-    # Expecting to see Linux methods prioritized, and then by method name
-    assert sort_methods_by_priority(
-        {WindowsA, WindowsB, WindowsC, LinuxA, LinuxB, LinuxC, MultiPlatformA}
-    ) == [LinuxA, LinuxB, LinuxC, MultiPlatformA, WindowsA, WindowsB, WindowsC]
-
-    monkeypatch.setattr("wakepy.core.method.CURRENT_PLATFORM", PlatformName.WINDOWS)
-    # Expecting to see windows methods prioritized, and then by method name
-    assert sort_methods_by_priority(
-        {WindowsA, WindowsB, WindowsC, LinuxA, LinuxB, LinuxC, MultiPlatformA}
-    ) == [MultiPlatformA, WindowsA, WindowsB, WindowsC, LinuxA, LinuxB, LinuxC]
-
-
-@pytest.mark.usefixtures("provide_methods_different_platforms")
-def test_get_prioritized_methods(monkeypatch):
-    WindowsA, WindowsB, WindowsC, LinuxA, LinuxB, LinuxC, MultiPlatformA = get_methods(
-        ["WinA", "WinB", "WinC", "LinuxA", "LinuxB", "LinuxC", "multiA"]
-    )
-
-    monkeypatch.setattr("wakepy.core.method.CURRENT_PLATFORM", PlatformName.LINUX)
-
-    assert get_prioritized_methods(
-        [
-            LinuxA,
-            LinuxB,
-            LinuxC,
-            MultiPlatformA,
-        ],
-        # Means: Prioritize LinuxC after everything else
-        methods_priority=["*", {"LinuxC"}],
-    ) == [LinuxA, LinuxB, MultiPlatformA, LinuxC]
-
-    assert get_prioritized_methods(
-        [
-            LinuxA,
-            LinuxB,
-            LinuxC,
-            MultiPlatformA,
-        ],
-        # Means: prioritize LinuxC over anything else.
-        methods_priority=[{"LinuxC"}],
-    ) == [LinuxC, LinuxA, LinuxB, MultiPlatformA]
-
-    assert get_prioritized_methods(
-        [
-            LinuxA,
-            LinuxB,
-            LinuxC,
-            MultiPlatformA,
-        ],
-        # Means, LinuxB first, then anything that is in between, and give
-        # lowest priority to LinuxA and LinuxC
-        methods_priority=[{"LinuxB"}, "*", {"LinuxA", "LinuxC"}],
-    ) == [LinuxB, MultiPlatformA, LinuxA, LinuxC]
-
-    assert get_prioritized_methods(
-        [LinuxA, LinuxB, LinuxC, MultiPlatformA, WindowsA],
-        # Means "LinuxB & WinA" first, ordered with automatic ordering, and
-        # then all the rest, also automatically ordered
-        methods_priority=[{"WinA", "LinuxB"}, "*"],
-    ) == [LinuxB, WindowsA, LinuxA, LinuxC, MultiPlatformA]
-
-    # No user-defined order -> Just alphabetical, but current platform (linux) first.
-    assert get_prioritized_methods(
-        [
-            LinuxA,
-            LinuxB,
-            WindowsA,
-            WindowsB,
-            LinuxC,
-            MultiPlatformA,
-        ],
-    ) == [LinuxA, LinuxB, LinuxC, MultiPlatformA, WindowsA, WindowsB]
-
-    monkeypatch.setattr("wakepy.core.method.CURRENT_PLATFORM", PlatformName.WINDOWS)
-    # No user-defined order -> Just alphabetical, but current platform (Windows) first.
-    assert get_prioritized_methods(
-        [LinuxA, LinuxB, WindowsA, WindowsB, LinuxC, MultiPlatformA],
-    ) == [MultiPlatformA, WindowsA, WindowsB, LinuxA, LinuxB, LinuxC]
