+++ wakepy/core/activation.py
@@ -21,6 +21,7 @@ import datetime as dt
 import os
 import typing
 from dataclasses import dataclass
+from typing import List, Set, Union
 
 from .constants import PlatformName
 from .method import MethodError, MethodOutcome
@@ -31,7 +32,10 @@ if typing.TYPE_CHECKING:
     from typing import Optional, Sequence, Tuple
 
     from .activationmanager import ModeActivationManager
-    from .method import Method
+    from .method import Method, MethodCls
+
+
+MethodsPriorityOrder = List[Union[str, Set[str]]]
 
 
 class UsageStatus(StrEnum):
@@ -246,6 +250,225 @@ class Heartbeat:
 
     def start(self):
         ...
+
+
+def check_methods_priority(
+    methods_priority: Optional[MethodsPriorityOrder], methods: List[MethodCls]
+) -> None:
+    """Checks against `methods` that the `methods_priority` is valid.
+
+    Parameters
+    ----------
+    methods_priority: list[str | set[str]]
+        The priority order, which is a list of where items are method names,
+        sets of methods names or the asterisk ('*'). The asterisk means "all
+        rest methods" and may occur only once in the priority order, and cannot
+        be part of a set. All method names must be unique and must be part of
+        the `methods`.
+    methods: list[MethodCls]
+        The methods which the `methods_priority` is validated against.
+
+    Raises
+    ------
+    ValueError or TypeError if the `methods_priority` is not valid.
+    """
+    if methods_priority is None:
+        return
+
+    known_method_names = {m.name for m in methods}
+    known_method_names.add("*")
+    seen_method_names = set()
+
+    for method_name, in_set in _iterate_methods_priority(methods_priority):
+        if not isinstance(method_name, str):
+            raise TypeError("methods_priority must be a list[str | set[str]]!")
+
+        if in_set and method_name == "*":
+            raise ValueError(
+                "Asterisk (*) may not be a part of a set in methods_priority!"
+            )
+        if method_name not in known_method_names:
+            raise ValueError(
+                f'Method "{method_name}" in methods_priority not in selected methods!'
+            )
+        if method_name in seen_method_names:
+            if method_name != "*":
+                raise ValueError(
+                    f'Duplicate method name "{method_name}" in methods_priority'
+                )
+            else:
+                raise ValueError(
+                    "The asterisk (*) can only occur once in methods_priority!"
+                )
+        seen_method_names.add(method_name)
+
+
+def _iterate_methods_priority(
+    methods_priority: Optional[MethodsPriorityOrder],
+) -> typing.Iterator[Tuple[str, bool]]:
+    """Provides an iterator over the items in methods_priority. The items in the
+    iterator are (method_name, in_set) 2-tuples, where the method_name is the
+    method name (str) and the in_set is a boolean which is True if the returned
+    method_name is part of a set and False otherwise."""
+
+    if not methods_priority:
+        return
+
+    for item in methods_priority:
+        if isinstance(item, set):
+            for method_name in item:
+                yield method_name, True
+        else:
+            yield item, False
+
+
+def get_prioritized_methods_groups(
+    methods: List[MethodCls], methods_priority: Optional[MethodsPriorityOrder]
+) -> List[Set[MethodCls]]:
+    """Prioritizes Methods in `methods` based on priority order defined by
+    `methods_priority`. This function does not validate the methods_priority in
+    any way; use `check_methods_priority` for validation of needed.
+
+    Parameters
+    ----------
+    methods: list[MethodCls]
+        The source list of methods. These methods are returned as prioritized
+        groups.
+    methods_priority: list[str | set[str]]
+        The names of the methods in `methods`. This specifies the priority
+        order; the order of method classes in the returned list. An asterisk
+        ('*') can be used to denote "all other methods".
+
+
+    Returns
+    -------
+    method_groups: list[set[MethodCls]]
+        The prioritized methods. Each set in the output represents a group of
+        equal priority. All Methods from the input `methods` are always
+        included in the output
+
+
+    Example
+    -------
+    Say there are methods MethodA, MethodB, MethodC, MethodD, MethodE, MethodF
+    with names "A", "B", "C", "D", "E", "F":
+
+    >>> methods = [MethodA, MethodB, MethodC, MethodD, MethodE, MethodF]
+    >>> get_prioritized_methods_groups(methods, methods_priority=["A", "F", "*"])
+    [
+        {MethodA},
+        {MethodF},
+        {MethodB, MethodC, MethodD, MethodE},
+    ]
+
+    """
+
+    # Make this a list of sets just to make things simpler
+    methods_priority_sets: List[Set[str]] = [
+        {item} if isinstance(item, str) else item for item in methods_priority or []
+    ]
+
+    method_dct = {m.name: m for m in methods}
+
+    asterisk = {"*"}
+    asterisk_index = None
+    out: List[Set[MethodCls]] = []
+
+    for item in methods_priority_sets:
+        if item == asterisk:
+            # Save the location where to add the rest of the methods ('*')
+            asterisk_index = len(out)
+        elif isinstance(item, set):
+            out.append({method_dct[name] for name in item})
+
+    out_flattened = {m for group in out for m in group}
+    rest_of_the_methods = {m for m in methods if m not in out_flattened}
+
+    if rest_of_the_methods:
+        if asterisk_index is not None:
+            out.insert(asterisk_index, rest_of_the_methods)
+        else:
+            out.append(rest_of_the_methods)
+
+    return out
+
+
+def sort_methods_by_priority(methods: Set[MethodCls]) -> List[MethodCls]:
+    """Sorts Method classes by priority and returns a new sorted list with
+    Methods with highest priority first.
+
+    The logic is:
+    (1) Any Methods supporting the CURRENT_PLATFORM are placed before any other
+        Methods (the others are not expected to work at all)
+    (2) Sort alphabetically by Method name, ignoring the case
+    """
+    return sorted(
+        methods,
+        key=lambda m: (
+            # Prioritize methods supporting CURRENT_PLATFORM over any others
+            0 if CURRENT_PLATFORM in m.supported_platforms else 1,
+            m.name.lower() if m.name else "",
+        ),
+    )
+
+
+def get_prioritized_methods(
+    methods: List[MethodCls],
+    methods_priority: Optional[MethodsPriorityOrder] = None,
+) -> List[MethodCls]:
+    """Take an unordered list of Methods and sort them by priority using the
+    methods_priority and automatic ordering. The methods_priority is used to define
+    groups of priority (sets of methods). The automatic ordering part is used
+    to order the methods *within* each priority group. In particular, all
+    methods supported by the current platform are placed first, and all
+    supported methods are then ordered alphabetically (ignoring case).
+
+    Parameters
+    ----------
+    methods:
+        The list of Methods to sort.
+    methods_priority:
+        Optional priority order, which is a list of method names (strings) or
+        sets of method names (sets of strings). An asterisk ('*') may be used
+        for "all the rest methods". None is same as ['*'].
+
+    Returns
+    -------
+    sorted_methods:
+        The input `methods` sorted with priority, highest priority first.
+
+    Example
+    -------
+    Assuming: Current platform is Linux.
+
+    >>> methods = [LinuxA, LinuxB, LinuxC, MultiPlatformA, WindowsA]
+    >>> get_prioritized_methods(
+    >>>    methods,
+    >>>    methods_priority=[{"WinA", "LinuxB"}, "*"],
+    >>> )
+    [LinuxB, WindowsA, LinuxA, LinuxC, MultiPlatformA]
+
+    Explanation:
+
+    WindowsA and LinuxB were given high priority, but since the current
+    platform is Linux, LinuxB was prioritized to be before WindowsA.
+
+    The asterisk ('*') is converted to a set of rest of the methods:
+    {"LinuxA", "LinuxC", "MultiPlatformA"}, and those are then
+    automatically ordered. As all of them support Linux, the result is
+    just the methods sorted alphabetically. The asterisk in the end is
+    optional; it is added to the end of `methods_priority` if missing.
+
+    """
+    unordered_groups: List[Set[MethodCls]] = get_prioritized_methods_groups(
+        methods, methods_priority=methods_priority
+    )
+
+    ordered_groups: List[List[MethodCls]] = [
+        sort_methods_by_priority(group) for group in unordered_groups
+    ]
+
+    return [method for group in ordered_groups for method in group]
 
 
 def activate_using(method: Method) -> MethodUsageResult:
